Add custom exception, add filtering pattern, update unit test with mock

+++ src/main/java/fr/viveris/s1pdgs/ingestor/services/file/FileProcessor.java
@@ -1,8 +1,6 @@
 package fr.viveris.s1pdgs.ingestor.services.file;
 
 import java.io.File;
-import java.util.concurrent.CancellationException;
-import java.util.concurrent.ExecutionException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,13 +9,17 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.messaging.Message;
 import org.springframework.stereotype.Component;
 
-import com.amazonaws.AmazonClientException;
-
 import fr.viveris.s1pdgs.ingestor.model.ConfigFileDescriptor;
 import fr.viveris.s1pdgs.ingestor.model.ErdsSessionFileDescriptor;
 import fr.viveris.s1pdgs.ingestor.model.ErdsSessionFileType;
 import fr.viveris.s1pdgs.ingestor.model.dto.KafkaMetadataDto;
 import fr.viveris.s1pdgs.ingestor.model.dto.KafkaSessionDto;
+import fr.viveris.s1pdgs.ingestor.model.exception.FilePathException;
+import fr.viveris.s1pdgs.ingestor.model.exception.IgnoredFileException;
+import fr.viveris.s1pdgs.ingestor.model.exception.KafkaMetadataPublicationException;
+import fr.viveris.s1pdgs.ingestor.model.exception.KafkaSessionPublicationException;
+import fr.viveris.s1pdgs.ingestor.model.exception.MetadataExtractionException;
+import fr.viveris.s1pdgs.ingestor.model.exception.ObjectStorageException;
 import fr.viveris.s1pdgs.ingestor.services.kafka.KafkaMetadataProducer;
 import fr.viveris.s1pdgs.ingestor.services.kafka.KafkaSessionProducer;
 import fr.viveris.s1pdgs.ingestor.services.s3.ConfigFilesS3Services;
@@ -64,26 +66,19 @@ public class FileProcessor {
 	/**
 	 * Builder of file descriptors
 	 */
-	@Autowired
 	FileDescriptorBuilder fileDescriptorBuilder;
 
 	/**
 	 * Builder of metadata
 	 */
-	@Autowired
 	MetadataBuilder metadataBuilder;
 
-	/**
-	 * Local directory when ERDS session files are temporarly stored
-	 */
-	@Value("${file.session-files.local-directory}")
-	public String sessionLocalDirectory;
-
-	/**
-	 * Local directory when configuration files are temporarly stored
-	 */
-	@Value("${file.config-files.local-directory}")
-	public String configLocalDirectory;
+	@Autowired
+	public FileProcessor(@Value("${file.config-files.local-directory}") final String configLocalDirectory,
+			@Value("${file.session-files.local-directory}") final String sessionLocalDirectory) {
+		this.metadataBuilder = new MetadataBuilder();
+		this.fileDescriptorBuilder = new FileDescriptorBuilder(configLocalDirectory, sessionLocalDirectory);
+	}
 
 	/**
 	 * Process configuration files.
@@ -101,49 +96,42 @@ public class FileProcessor {
 		}
 		// Build model file
 		try {
-			ConfigFileDescriptor descriptor = fileDescriptorBuilder.buildConfigFileDescriptor(file);
-			// Store in object storage
-			if (descriptor.isHasToBeStored()) {
-				configFilesS3Services.uploadFile(descriptor.getRelativePath(), file);
-			}
-			// Extract metadata
-			if (descriptor.isHasToExtractMetadata()) {
-				KafkaMetadataDto extractedMetadata = metadataBuilder.buildConfigFileMetadata(descriptor, file);
-				try {
+			try {
+				ConfigFileDescriptor descriptor = fileDescriptorBuilder.buildConfigFileDescriptor(file);
+				// Store in object storage
+				if (descriptor.isHasToBeStored()) {
+					if (!configFilesS3Services.exist(descriptor.getRelativePath())) {
+						configFilesS3Services.uploadFile(descriptor.getRelativePath(), file);
+						LOGGER.info("[processConfigFile] {} successfully published is object storage",
+								descriptor.getRelativePath());
+					} else {
+						throw new IgnoredFileException(descriptor.getProductName(),
+								new Exception("File already exist in object storage"));
+					}
+				}
+				// Extract metadata
+				if (descriptor.isHasToExtractMetadata()) {
+					KafkaMetadataDto extractedMetadata = metadataBuilder.buildConfigFileMetadata(descriptor, file);
 					senderMetadata.send(extractedMetadata);
 					LOGGER.info("[processConfigFile] Metadata for {} successfully sended",
 							descriptor.getRelativePath());
-				} catch (CancellationException | InterruptedException | ExecutionException e) {
-					LOGGER.error("[processConfigFile] Metadata not published in message queuing for {}",
-							descriptor.getRelativePath());
+				}
+
+			} catch (IgnoredFileException | MetadataExtractionException | FilePathException
+					| KafkaMetadataPublicationException ce) {
+				LOGGER.error(ce.getMessage());
+			} finally {
+				// Remove file
+				if (!file.isDirectory()) {
+					if (!file.delete()) {
+						LOGGER.error("[processConfigFile] File {} not removed from local storage", file.getPath());
+					}
 				}
 			}
-			// Remove file
-			LOGGER.info("[processConfigFile] Try for {} remove file",
-					descriptor.getRelativePath());
-			if (!file.isDirectory()) {
-				if (!file.delete()) {
-					LOGGER.error("[processConfigFile] File {} not removed from local storage",
-							descriptor.getRelativePath());
-				} else {
-					LOGGER.info("[processConfigFile] File {} successfully deleted",
-							descriptor.getRelativePath());
-				}
-			} else {
-				LOGGER.info("[processConfigFile] File {} is a directory, must not remove",
-						descriptor.getRelativePath());
-			}
-		} catch (AmazonClientException ae) {
-			LOGGER.error("Processing of configuration file {} failed (step: object storage): {}", file.getPath(),
-					ae.getMessage());
-		} catch (IllegalArgumentException iae) {
-			LOGGER.error("Processing of configuration file {} bypassed: {}", iae.getMessage());
-		} catch (Exception e) {
-			LOGGER.error("Processing of configuration file {} failed: {}", e.getMessage());
+		} catch (ObjectStorageException fe) {
+			LOGGER.error(fe.getMessage());
 		}
-		if (LOGGER.isDebugEnabled()) {
-			LOGGER.debug("Processing of configuration file {} succeeded", file.getPath());
-		}
+
 		if (LOGGER.isInfoEnabled()) {
 			LOGGER.info("End processing of configuration file {}", file.getPath());
 		}
@@ -166,45 +154,42 @@ public class FileProcessor {
 		}
 		// Build model file
 		try {
-			ErdsSessionFileDescriptor descriptor = fileDescriptorBuilder.buildErdsSessionFileDescriptor(file);
-			// Store in object storage
-			sessionFilesS3Services.uploadFile(descriptor.getRelativePath(), file);
-			// Extract metadata
-			KafkaMetadataDto extractedMetadata = metadataBuilder.buildErdsSessionFileMetadata(descriptor, file);
 			try {
+				ErdsSessionFileDescriptor descriptor = fileDescriptorBuilder.buildErdsSessionFileDescriptor(file);
+				// Store in object storage
+				if (!sessionFilesS3Services.exist(descriptor.getRelativePath())) {
+					sessionFilesS3Services.uploadFile(descriptor.getRelativePath(), file);
+				} else {
+					throw new IgnoredFileException(descriptor.getProductName(),
+							new Exception("File already exist in object storage"));
+				}
+				// Extract metadata
+				KafkaMetadataDto extractedMetadata = metadataBuilder.buildErdsSessionFileMetadata(descriptor, file);
 				senderMetadata.send(extractedMetadata);
-			} catch (CancellationException | InterruptedException | ExecutionException e) {
-				LOGGER.error("[processConfigFile] Metadata not published in message queuing for {}",
-						descriptor.getRelativePath());
-			}
-			// Publish session file
-			if (descriptor.getProductType() == ErdsSessionFileType.SESSION) {
-				try {
-					// TODO use a transformer
+
+				// Publish session file
+				if (descriptor.getProductType() == ErdsSessionFileType.SESSION) {
 					KafkaSessionDto dtoSession = new KafkaSessionDto();
 					dtoSession.setProductName(descriptor.getProductName());
 					dtoSession.setKeyObjectStorage(descriptor.getKeyObjectStorage());
 					dtoSession.setChannel(descriptor.getChannel());
 					dtoSession.setSessionIdentifier(descriptor.getSessionIdentifier());
 					senderSession.send(dtoSession);
-				} catch (CancellationException | InterruptedException | ExecutionException e) {
-					LOGGER.error("[processConfigFile] Session not published in message queuing for {}",
-							descriptor.getRelativePath());
+				}
+
+			} catch (IgnoredFileException | MetadataExtractionException | FilePathException
+					| KafkaSessionPublicationException | KafkaMetadataPublicationException ce) {
+				LOGGER.error(ce.getMessage());
+			} finally {
+				// Remove file
+				if (!file.isDirectory()) {
+					if (!file.delete()) {
+						LOGGER.error("[processConfigFile] File {} not removed from local storage", file.getPath());
+					}
 				}
 			}
-			// Remove file
-			if (!file.delete()) {
-				LOGGER.error("[processConfigFile] File {} not removed from local storage",
-						descriptor.getRelativePath());
-			}
-			if (LOGGER.isDebugEnabled()) {
-				LOGGER.debug("Processing of ERDS file {} succeeded", file.getPath());
-			}
-		} catch (AmazonClientException ae) {
-			LOGGER.error("Processing of ERDS session file {} failed (step: object storage): {}", file.getPath(),
-					ae.getMessage());
-		} catch (IllegalArgumentException iae) {
-			LOGGER.error("Processing of file {} bypassed: {}", iae.getMessage());
+		} catch (ObjectStorageException fe) {
+			LOGGER.error(fe.getMessage());
 		}
 		if (LOGGER.isInfoEnabled()) {
 			LOGGER.info("End processing of ERDS session file {}", file.getPath());
